+++ crates/anodized-core/src/instrument/traits/tests.rs
@@ -1,0 +1,60 @@
+use crate::test_util::assert_tokens_eq;
+
+use super::*;
+use proc_macro2::TokenStream;
+use syn::{ItemTrait, parse_quote};
+
+#[test]
+fn embed_spec_item_trait() {
+    let trait_spec = Spec::empty();
+    let item_trait: ItemTrait = parse_quote! {
+        trait TRAIT {
+            #[spec(
+                requires: COND_1,
+                maintains: COND_2,
+                binds: PAT_1,
+                ensures: COND_3,
+            )]
+            fn FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) -> RET_TYPE {
+                BODY
+            }
+        }
+    };
+
+    let expected: TokenStream = parse_quote! {
+        trait TRAIT {
+            fn FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) -> RET_TYPE {
+                Self::__anodized_FUNC(self, PARAM_1, PARAM_2)
+            }
+
+            #[doc(hidden)]
+            fn __anodized_FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) -> RET_TYPE {
+                BODY
+            }
+
+            #[doc(hidden)]
+            #[allow(warnings)]
+            fn __anodized_fn_requires_FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) {
+                let _ = | | COND_1;
+            }
+
+            #[doc(hidden)]
+            #[allow(warnings)]
+            fn __anodized_fn_maintains_FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) {
+                let _ = | | COND_2;
+            }
+
+            #[doc(hidden)]
+            #[allow(warnings)]
+            fn __anodized_fn_ensures_FUNC(&self, PARAM_1: TYPE_1, PARAM_2: TYPE_2) {
+                let () = ();
+                let _ = |PAT_1: &RET_TYPE| COND_3;
+            }
+        }
+    };
+
+    let observed = Backend::NOTHING
+        .instrument_item_trait(trait_spec, item_trait)
+        .unwrap();
+    assert_tokens_eq(&observed, &expected);
+}
